+++ src/app/signup/SignupForm.tsx
@@ -4,7 +4,8 @@ import Input from "@/app/_component/common/Input";
 import { cn } from "@/utils/cn";
 import { useEffect, useState } from "react";
 import { useForm } from "react-hook-form";
-import { useSignUp, useIdDuplicateCheck } from "../hooks/queries/useAuth";
+import { useSignUp } from "../hooks/mutations/useSignup";
+import { useIdDuplicateCheck } from "../hooks/mutations/useIdDuplicateCheck";
 import Toast from "../_component/common/Toast";
 
 type LoginFormValues = {
