+++ src/app/types/loginForm.ts
@@ -1,0 +1,4 @@
+export interface LoginFormValues {
+  email: string;
+  password: string;
+}
